# openapi.yaml
openapi: 3.0.0
info:
  title: Status MCP API
  description: API for managing user status context
  version: 1.0.0
servers:
  - url: http://localhost:3000
paths:
  /status:
    get:
      summary: Get current status
      responses:
        '200':
          description: Resolved status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
    put:
      summary: Set manual status override
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                reason:
                  type: string
                ttlSeconds:
                  type: integer
      responses:
        '200':
          description: Updated status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
  /status/date/{date}:
    get:
      summary: Get status for a specific date
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Resolved status for date
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
  /status/schedule:
    get:
      summary: Get scheduled overrides
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: List of schedules
    put:
      summary: Set scheduled override
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                date:
                  type: string
                  format: date
                workStatus:
                  type: string
                reason:
                  type: string
      responses:
        '200':
          description: Success
components:
  schemas:
    Status:
      type: object
      properties:
        effectiveDate:
          type: string
        resolvedAt:
          type: string
        bankHoliday:
          type: boolean
        weekend:
          type: boolean
        workStatus:
          type: string
        location:
          $ref: '#/components/schemas/Location'
        lastUpdated:
          type: string
    Location:
      type: object
      properties:
        latitude:
          type: number
        longitude:
          type: number
        locationName:
          type: string
        source:
          type: string
        timestamp:
          type: string